Parser

TotalViewITCH.Parser — Type

Parser

A struct used to parse TotalViewITCH files and their contents to disk.

Details

Parsers are responsible for deserializing order book messages stored in raw (binary) TotalViewITCH files and generating order book snapshots. Parsers wrap Backend objects that write messages and order book snapshots to disk. By default, parsers maintain the first five levels of each order book the user requests to track and writes out data after every buffer_size messages read for each ticker (separately).

Examples

parser = Parser{FileSystem}("./data/test")
parser("./data/bin/S031413-v41.txt", Date("2013-03-14"), ["A"], 4.1)

parser = Parser{MongoDB}("mongodb://localhost:27017")
parser("data/bin/S022717-v50.txt", Date("2017-02-27"), ["A"], 5.0)

Backends

TotalViewITCH.FileSystem — Type

FileSystem <: Backend

A backend for storing data to the local file system in CSV format.

Data is stored with the following directory structure:

root
|- collection
   |- ticker
      |- date
         |- partition.csv

TotalViewITCH.MongoDB — Type

MongoDB <: Backend

A backend for storing data to MongoDB.

The default database name is "totalview-itch", which contains the following collections:

messages

orderbooks

noii

trades

All collections are indexed by ticker and date fields.

TotalViewITCH.ping — Function

ping(b::Backend)

Check if the backend can connect to the database.

TotalViewITCH.build — Function

build(b::Backend; kwargs)

Scaffold a database. By default, the program prompts the user to overwrite existing files. Set force=true to overwrite existing files without prompting.

TotalViewITCH.clean — Function

clean(date::Date, ticker::String, b::Backend)

Remove all data found for the specified date and ticker.

clean(date::Date, b::Backend)

Remove all data found for the specified date.

clean(ticker::String, b::Backend)

Remove all data found for the specified ticker.

TotalViewITCH.teardown — Function

teardown(b::Backend)

Remove all database files. Set force=true to skip the default confirmation prompt.

TotalViewITCH.find — Function

find(b::Backend, collection, ticker, date)

Load processed data for the specified collection, ticker and date, if available.

Messages

TotalViewITCH.OrderMessage — Type

OrderMessage

A simple data type representing order book updates.

Order book updates communicate changes to the order book in response to trader actions, e.g., add, execute, delete, etc.

TotalViewITCH.SystemMessage — Type

SystemMessage

A simple data type representing system updates.

System updates comunicate changes that apply to the entire exchange, such as the beginning and ending of trading hours.

TotalViewITCH.NOIIMessage — Type

NOIIMessage

A simple data type representing net order imbalance indicator messages and cross trade messages.

TotalViewITCH.TradeMessage — Type

TradeMessage

A simple data type representing trade messages.

Order Books

TotalViewITCH.Order — Type

Order

A limit order.

TotalViewITCH.Book — Type

Book

A limit order book.

Arguments

ticker::String: the associated security ticker.

nlevels::Int: the number of levels reported in tabular format.