TotalViewITCH.process — Function

process(file, version, date, nlevels, tickers, path)

Read a binary data file and write message and order book data to file.

Arguments

file: location of file to read.

version: ITCH version number (4.1 or 5.0).

date: Date to associate with output.

nlevels: number of order book levels to track.

tickers: stock tickers to track.

dir: location to write output.

TotalViewITCH.Recorder — Type

Recorder

A data structure to manage writing data to CSV files.

The Recorder holds lines in a string. When the Recorder is full, the string is written to file and the string is emptied.

Note that push! adds an end-of-line character to line. Thus, lines pushed to Recorders should not include ' '.

TotalViewITCH.build — Function

build(dir)

Scaffold a database at dir. The structure of the database is:

dir
 |- books
     |- aapl.csv
     |- ...
 |- messages
     |- aapl.csv
     |- ...
 |- trades
     |- aapl.csv
     |- ...
 |- noii
     |- aapl.csv
     |- ...

TotalViewITCH.teardown — Function

teardown(dir; <kwargs>)

Delete a database at dir.

TotalViewITCH.OrderMessage — Type

OrderMessage

Data structure representing order book updates.

TotalViewITCH.Book — Type

Book

A limit order book.

Arguments

name::String: the associated security name/ticker

nlevels::Int: the number of levels reported

TotalViewITCH.Order — Type

Order

A limit order.

TotalViewITCH.TradeMessage — Type

TradeMessage

Data structure representing trades.

TotalViewITCH.SystemMessage — Type

SystemMessage

Data structure representing system updates.

TotalViewITCH.NOIIMessage — Type

NOIIMessage

Data structure representing net order imbalance indicator messages and cross trade messages.